remove and delete CIS in the webapp added missing implementation of a method in ICISsimulated on CPA test

+++ platform-infrastructure/web-app/src/main/webapp/WEB-INF/views/community_profile.jsp
@@ -67,8 +67,15 @@
 <section  class="grid_12">
 <section>
 <div class="breadcrumbs"><a href="">Home</a> / <a href="community_profile.html?cisId=${cisInfo.getCommunityJid()}">${cisInfo.getCommunityName()}</a></div>
+<br>
+<xc:if test="${response != 'null'}">
+	<div class="success">${response}</div>
+</xc:if>
 </section>
 <div class="websearchbar">
+
+
+
 <div class="websearchtitle">
 <h4 class="profile_title">${cisInfo.getCommunityName()} Profile</h4>
 </div>
@@ -84,7 +91,14 @@
 <section>
 <figure class="gravatar">
 <img alt="" src="images/webcommunity_pic_sample1.jpg" height="48" width="48" />
-<a class="furtherinfo-link" href="delete_community.html?cisId=${cisInfo.getCommunityJid()}">REMOVE</a>
+
+	<xc:if test="${isOwner == true}">
+		<a class="furtherinfo-link" href="delete_community.html?cisId=${cisInfo.getCommunityJid()}" onclick="return confirm('Are you sure you want to delete the CIS?')">REMOVE</a>
+	</xc:if>
+	<xc:if test="${isOwner == false}">
+		<a class="furtherinfo-link" href="leave_community.html?cisId=${cisInfo.getCommunityJid()}" onclick="return confirm('Are you sure you want to leave the CIS?')">LEAVE</a>
+	</xc:if>
+
 </figure>
 <div class="keyinfo_content">
 <div class="clearfix">
